requirements.txt
mcp>=0.1.0
pydantic>=2.0.0
pyyaml>=6.0
jinja2>=3.1.0
requests>=2.31.0

